added check for file existence before trying to delete it

+++ src/Illuminate/Filesystem/Filesystem.php
@@ -216,12 +216,14 @@ class Filesystem
         $success = true;
 
         foreach ($paths as $path) {
-            try {
-                if (! @unlink($path)) {
+            if (file_exists($path)) {
+                try {
+                    if (!@unlink($path)) {
+                        $success = false;
+                    }
+                } catch (ErrorException $e) {
                     $success = false;
                 }
-            } catch (ErrorException $e) {
-                $success = false;
             }
         }
